Centralise workspace actionsi n `WorkspaceManager`
This commit moves all methods from `WorkspaceSetup` and `CheckoutAction`
and places them into `WorkspaceManager`. Additionally, method
`CheckoutAction.checkout()` has been split into two separate methods:
* `WorkspaceManager.prepare()` to create and cleanup the workspace
* `WorkspaceManager.setSelector()` to set the selector

Alternatively, there's a `WorkspaceManager.switchTo()` method to be used
by `MergebotScm`.
